♻️ Update and Prepare for Workspaces
♻️ Update and Prepare for Workspaces

✨ Updates the project with the following changes:

- Rust edition bumped from `2018` to `2023`
- dependencies updated to latest versions
- addressed `cargo clippy` lint warnings

✨ Preparation for Workspaces

I want to begin separating the logic of this crate out
so that it's easier to adapt the following concepts:

- Integration into my `neovim` workflow
- A graphical CLI optional interface
- Integration into the mdBook project
- `curl` conversion
- Automated testing system

✨ Add HTTP Lib

Introduces a simple http model for both the request
and the response. Also provides a client interface
trait for reqwest. This is to make working with the
models more universal across different platforms.